backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.6.0
pydantic-settings==2.2.0
python-multipart==0.0.6
pillow==10.1.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
httpx==0.25.0
